backend/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
fastkml==1.4.0
gunicorn==25.0.3
minio==7.2.20
numpy==2.4.2
packaging==26.0
psycopg==3.3.2
psycopg-binary==3.3.2
pycparser==3.0
pycryptodome==3.23.0
pygeoif==1.6.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
shapely==2.1.2
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
